+++ common/src/main/cpp/register_types.cpp
@@ -64,6 +64,7 @@
 #include "classes/openxr_fb_passthrough_geometry.h"
 #include "classes/openxr_fb_render_model.h"
 #include "classes/openxr_fb_scene_manager.h"
+#include "classes/openxr_fb_spatial_anchor_manager.h"
 #include "classes/openxr_fb_spatial_entity.h"
 #include "classes/openxr_fb_spatial_entity_query.h"
 #include "classes/openxr_meta_passthrough_color_lut.h"
@@ -142,6 +143,7 @@ void initialize_plugin_module(ModuleInitializationLevel p_level) {
 			ClassDB::register_class<OpenXRFbRenderModel>();
 			ClassDB::register_class<OpenXRFbHandTrackingMesh>();
 			ClassDB::register_class<OpenXRFbSceneManager>();
+			ClassDB::register_class<OpenXRFbSpatialAnchorManager>();
 			ClassDB::register_class<OpenXRFbSpatialEntity>();
 			ClassDB::register_class<OpenXRFbSpatialEntityQuery>();
 			ClassDB::register_class<OpenXRFbPassthroughGeometry>();