+++ CHANGES.md
@@ -11,6 +11,7 @@
 - Add XR_FB_hand_tracking_aim support
 - Update Meta OpenXR mobile SDK to version 62
 - Add a developer-facing API for interacting with scene anchors
+- Add support for developer-created spatial anchors via XR_FB_spatial_entity
 - Add XR_FB_hand_tracking_capsules extension wrapper
 - Add OpenXRFbPassthroughGeometry node
 - Add OpenXRMetaPassthroughColorLut